chore: added more verification for change password queries
