+++ src/app/api/auth/change-password/route.ts
@@ -1,5 +1,30 @@
 import { changePassword, changePasswordQuery, isLegit } from "@/lib/query/users/changePassword"
+import { findChangePasswordQuery } from "@/lib/query/users/findChangePasswordQuery"
 import { ReasonPhrases, StatusCodes } from "http-status-codes"
+
+export async function GET(req: Request) {
+  const { searchParams } = new URL(req.url)
+  const token = searchParams.get("token")
+
+  if (!token) {
+    return Response.json({ message: "Missing token" }, { status: StatusCodes.BAD_REQUEST })
+  }
+
+  try {
+    const resetQuery = await findChangePasswordQuery(token)
+
+    if (!resetQuery) {
+      return Response.json({ message: "Invalid token" }, { status: StatusCodes.NOT_FOUND })
+    }
+
+    return Response.json({ message: "Valid token" }, { status: StatusCodes.OK })
+  } catch (error) {
+    return Response.json(
+      { message: ReasonPhrases.INTERNAL_SERVER_ERROR },
+      { status: StatusCodes.INTERNAL_SERVER_ERROR }
+    )
+  }
+}
 
 export async function POST(req: Request) {
   const { email, locale } = (await req.json()) as PostPayload
@@ -24,7 +49,7 @@ export async function POST(req: Request) {
 }
 
 export async function PATCH(req: Request) {
-  const { token, currentPassword, newPassword } = (await req.json()) as PutPayload
+  const { token, currentPassword, newPassword } = (await req.json()) as PatchPayload
 
   if (!token || !currentPassword || !newPassword) {
     return Response.json(
@@ -38,7 +63,7 @@ export async function PATCH(req: Request) {
   try {
     const userId = await isLegit(token, currentPassword)
 
-    await changePassword(userId, newPassword)
+    await changePassword(token, userId, newPassword)
 
     return Response.json({ message: "Password changed" }, { status: StatusCodes.OK })
   } catch (error) {
@@ -54,7 +79,7 @@ type PostPayload = {
   locale: string
 }
 
-type PutPayload = {
+type PatchPayload = {
   token: string
   currentPassword: string
   newPassword: string
