+++ src/lib/query/users/confirmChangePassword.ts
@@ -1,0 +1,15 @@
+import axios from "axios"
+
+export async function confirmChangePassword(
+  token: string,
+  currentPassword: string,
+  newPassword: string
+) {
+  const data = {
+    token,
+    currentPassword,
+    newPassword
+  }
+
+  await axios.patch(`/api/auth/change-password`, data)
+}
