+++ src/lib/query/users/findChangePasswordQuery.ts
@@ -3,7 +3,11 @@ import prisma from "@/lib/db"
 export async function findChangePasswordQuery(token: string) {
   return await prisma.passwordChange.findFirst({
     where: {
-      token
+      token,
+      expires: {
+        gte: new Date()
+      },
+      usedAt: null
     }
   })
 }
