+++ src/lib/query/users/findChangeToken.ts
@@ -2,7 +2,7 @@ import axios from "axios"
 
 export default async function findChangeToken(token: string) {
   try {
-    await axios.get(`/api/auth/find-change-query?token=${token}`)
+    await axios.get(`/api/auth/change-password?token=${token}`)
 
     return { tokenVerified: true }
   } catch (error) {
