+++ src/lib/utils/change-password/sendResetPasswordEmail.ts
@@ -22,8 +22,8 @@ export async function sendResetPasswordEmail(email: string, token: string, local
     to: email,
     subject: "Change your password",
     html: `
-      <p>Click the link below to reset your password:</p>
-      <a href="${url}/${locale}/change-password?token=${token}">Reset Password</a>
+      <p>Click the link below to change your password:</p>
+      <a href="${url}/${locale}/change-password?token=${token}">Change Password</a>
     `
   }
