+++ src/validators/schemas/changePasswordSchema.ts
@@ -1,7 +1,12 @@
 import { z } from "zod"
 
-export const changePasswordSchema = z.object({
-  currentPassword: z.string(),
-  newPassword: z.string().min(3),
-  confirmPassword: z.string().min(3)
-})
+export const changePasswordSchema = z
+  .object({
+    currentPassword: z.string(),
+    newPassword: z.string().min(3),
+    confirmPassword: z.string().min(3)
+  })
+  .refine((data) => data.newPassword === data.confirmPassword, {
+    message: "Passwords do not match",
+    path: ["confirmPassword"]
+  })
